Update to MC 1.16.2
The mod now works with Minecraft 1.16.2. The mod has been compiled with Forge 1.16.2-33.0.21.

Due to the early development stages of 1.16.2 Forge, ore/structure generation has not yet been implemented. The abandoned factory structure will not spawn naturally in the world, and ores will not generate. This functionality will be re-implemented once Forge updates.