Update 0.6.0
A major update, adding new items, cleaning up old files, a bit of refactoring, and implementing support for a popular mod, JEI.

To start, I've removed the -alpha1 suffix from this release. That's right, its out of alpha now. This means that there should be sufficient content added to really add something to your gameplay. It also means that I'll start releasing binaries in the Releases area of this repository. You can download these and use them with Forge as you would any other mod.

On to items. The main addition in this update was the Flintlock Pistol. It can fire multiple type of ammo, being created from different material tiers. (iron, gold, diamond, even netherite, etc).
The pistol uses musket balls as ammo. Be aware, if you use low quality ammo (wood/stone musket balls), your pistol might misfire!
If you're using the pistol in Creative mode with no ammo, it'll default to Iron musket balls.

Added a new crafting machine. This machine is called the "Small Parts Table". As the name suggests, it's used to make small pieces (like those used in the Flintlock Pistol recipe). It uses material items along with a blueprint to create the part(s) you need. Bonus: your blueprint item won't be consumed, so you can reuse it infinitely! Along with this, there's also a custom use sound when you craft something inside it.

Going along with the new crafting machine, I've added support for a very popular mod, Just Enough Items (JEI). This means you can use JEI to view recipes for the Small Parts Table and quick transfer ingredients. Of course, you don't need JEI to run the mod. But if you do use it, you'll see support there.

The few other changes are the movement from static items to tags in some ingredients, to allow a wider variety of items to satistfy recipes and the minor refactoring of packages.

The last thing to note is the update from Minecraft 1.16.4 to 1.16.5. Don't panic! You can use the mod in both 1.16.4 AND 1.16.5.