New traps and bugfixes
New traps! First is the landmine. Pretty simple, you step on it, you explode. You can disarm landmines with pliers.

The second is barbed wire. Not much of a trap, but can be used to defend your base. Like sweet berries but more aggressive.

Traps can now randomly generate in different biomes. Landmines in deserts, bear traps in forests, and pitfalls in the jungle. You can configure the separation distances in the configuration file.

You can also camouflage the bear traps and landmine. The bear trap can be camouflaged with vines, while the landmine can be camouflaged with vines and sand.

Other changes include some bug fixes with other traps, and a bit of minor refactoring (mostly due to switching from Eclipse to IntelliJ)