Create new block and minor balancing
New block! Introducing the wall shelf, so you can decorate your base a bit! You can place it on walls, then right-click with an item in your hand to place it on the shelf. Right-click again with an empty hand to remove the most recently added item.

In terms of minor balancing, I reduced the amount of concrete needed to craft the spotlight. It now only requires two black concrete blocks instead of the previous four blocks.