New biome, entities, block, worldgen, optimizations, etc.
I could easily describe this commit as everything, because I've worked on most aspects of the mod. The main addition with this commit is the new battlefield biome, a bloodred zone where enemy soldiers can be found (and minutemen protecting their villages).

The new entity is the minuteman. Similar to the dying soldier, it wields a blunderbuss and has similar attack patterns. They work in groups and protect villages like iron golems do.

The new block is a wooden spike wall that acts like barbed wire, only you can't pass through it and it deals less damage. It also doesn't slow you down much.

For worldgen, I have written a new world carver to be used in the battlefield biome. It's a modified version of the ravine carver, designed to form trenches.

In terms of optimizations, I've updated Forge to the latest version, and allowed IntelliJ to reorganize and optimize files. I've also went and removed unnecessary imports from most files.