Create bandages and MREs
This commit adds two new items. The first is a bandage, which upon use will provide the player with 12 seconds of regeneration 1 (healing 2.5 hearts). It can be used on other entities to provide the effect to them as well.

The second item is an MRE, a food item providing 11 hunger points and 12.6 saturation. It is created from four foods (carrot, potato, beetroot, and cooked chicken) in a shapeless recipe.

I also moved some mod events from the main mod file to the ModEventSubscriber file. I also removed the long-unused empty bullet casing models and texture.