Add custom spawn eggs to vanilla SpawnEggItem.BY_ID map
This change uses reflection to access the vanilla SpawnEggItem.BY_ID map
and add the mod spawn eggs to that map. This means that players in
creative mode will now be able to middle click the mod entities to get
the spawn eggs (like the vanilla entities). This also means that other
mods will be able to query the BY_ID map to get the spawn eggs from this
mod.