Add mud blocks and various subtypes
Add mud blocks, and subtypes of dried mud and hardened mud. Mud blocks have a 10% chance to "dry" and convert to dried mud every random tick, but if it is raining it is a 3% chance. Conversely, dried mud blocks have a 10% chance to convert to mud every random tick if it is touching a water source. Rain increases this to 13%.

Mud blocks slow down entities by 25% (excluding dried and hardened types).

Dried mud can be smelted into hardened mud, which is not affected by nearby water sources. Hardened mud also has slab and stair variants.

Fix a bug regarding blockstates of burned oak slabs where the top half of a double slab would be invisible.

Update Forge to 37.0.103, and Parchment mappings to the latest version.