Initial 1.18 Port
This commit covers the initial porting steps for 1.18.

Everything relating to worldgen is disabled, due to problems regarding 1.18 and the Forge API. It will be reenabled sometime in the future.

JEI support is disabled until a 1.18 release is available for JEI.

Update Forge from version 37.1.0 to 38.0.12, WTHIT from 3.11.2 to 4.3.0, and change mappings from Parchment 1.17 to Official 1.18 until a Parchment 1.18 release is available.