Add bullet impact particles and revamp musket ball models
Impact particles like the sandbag impact particle now exist on any block that is shot with a musket ball. The sandbag-specific functionality has been removed now that the bullet applies it to everything.

Rework musket ball models, for both the inventory and when fired into the world.

Make muzzle flash particle change based on player FOV.

Fix mods.toml pointing to dev branch instead of master for the update json.

A bit off scope, but Lava Revenants can now only destroy blocks with a resistance of 1.5 or less.