Create musket item
Create a musket item. It is a long-ranged firearm with a scope attached. The range/accuracy of the flintlock pistol and blunderbuss has been reduced for balance. There is no crafting recipe available yet for the musket.

Landmines now explode when shot by a projectile.

Update mappings from 2022.03.13 to 2022.05.02, Forge from 40.1.0 to 40.1.16, and WTHIT from 4.9.1 to 4.10.1.