Create musket variant and gun model rework
Added a scope-less version of the musket.

Reworked models for the flintlock pistol, blunderbuss, and flare gun.

Added custom firing sound for the musket.

Prevent scoping in with the musket when the player is not in first person. Also fixed a visual bug regarding the overlay, particularly at night.

Removed one of the custom wood textures and replaced it with a simple brown for cleanliness.

Updated Forge from 40.1.16 to 40.1.19.