Tweak entity spawning
Configuration options are back in the `.toml` file for entity spawns. Options are available for enabling/disabling spawns, setting spawn weights, and settings min/max spawn clusters.

Spawning balances have been changed. Rejoice, as Wandering Warriors should no longer spawn everywhere. Additionally, more spawn rules have been added to most entities. For entities that stick to the ground, these rules check that the entity has a valid walk target, is not in water, is on ground, and if it can see the sky. This prevents mass spawning underground, particularly in the cases of the Wandering Warrior and Dying Soldiers. For entities that fly, the spawn rules check to see if the entity is in the air and not in water, and there is a sufficient area of empty space (air blocks) around them.