Rename some old advancements internally
Change the names of some old advancements which have been tagged for renaming since February... This is mostly internal and just makes it easier to determine what the advancements are for when looking at the filenames. The translation keys still reflect the original names.

Update TODO in `BlockStateGenerator` with new PR number.