New endgame item tier
Add Starstorm, a new endgame tier in Tiltros. It specializes in raw damage output and high attack speed (not quite as fast as Astral). Wearing a full set of Starstorm armor will grant a 20% damage increase to all outgoing attacks.

Starstorm ingots can be crafted by crushing Starstorm Crystals with pistons from above. These crystals are found in Celestial Asteroids in the Deadman's Desert. When crushing crystals, there is a 15% chance to spawn a Starmite, a creature similar to the Endermite.

Minor rebalancing of existing tiers.

Include Target Dummy as a runtime only dependency for testing damage output on weapons.