Buff molten armor
Molten previously had one of the weakest setups in terms of attack prowess compared to other endgame tiers. Now, it has a 20% damage bonus when in the Nether, plus another 10% when submerged in lava. This will be a strong benefit to those who take the extra challenge of building Nether bases.