New structure
Add the Biodome structure, which spawns in the Deadman's Desert. It contains some survival gear for those wandering in this biome. Inside is also a new block, the Biodome Life Support Unit. When powered by redstone, it will grant nearby entities Celestial Protection. It is uncraftable, so the only way to obtain them is to find these structures.

Fix broken spotlight blockstate, which was missing rotation values.

Reorganize structure template pool files.