Add meteor staff item
Add a meteor staff item, which summons meteors from the sky when right-clicking. Players can point to where it should fall to, and it has a default max range of 100 blocks out.

The max use range, blast radius, and block breaking can all be configured.

Fix bugs in the configuration handler that would cause some values to reset to default.