Add Storm Creeper entity
As the name implies the Storm Creeper is a variant of the overworld Creeper. It has a higher blast radius and is 25% larger in size. They share the same HP and defense, however the storm variant is significantly slower. It will spawn naturally in the Deadman's Desert.

Make the Meteor Staff repairable.

Properly set the new biome tracks as music instead of ambient.