Add new structure and NPC
Add the Skygazer NPC, a villager-like entity which provides endgame level trades. It is found in Space Observatory structures which generate randomly in Starlight Plains biomes.

Add language entries for enchantment levels 11-25 as the Skygazer can theoretically increase a level past 10 (the max numeral vanilla has).

Add a rusted iron block for decoration purposes. There is no crafting recipe and it is not used in anything else.

Water towers and underground bunkers will generate with some of their iron blocks rusted, which makes it much harder to dismantle them for easy sources of iron.

Adjust the spawn rates of some entities.