Add some enchantments
Add some enchantments to guns. So far, this includes three enchants that are equivalent those on bows (power, punch, and infinity).

Infinity on bows (and the equivalent "Endless Musket Pouch" on guns) is now restricted to cobalt-tier ammo or lower, for balance. This can be overridden with a config option.