More gun enchants
Add three new enchantments, Scorch Shot, Rapid Fire, and Velocity. The first is an equivalent to the vanilla Flame enchant for guns.

Rapid fire reduces reload time on guns by 5% per level up with a maximum of 3 levels.

Velocity can be applied to bows and guns. It increases the velocity of projectiles by 10% per level with a maximum of 3 levels.