Another batch of new enchantments
Add five new enchantments to various weapons.

Crimson Claw is a gauntlet enchantment which increases the bleed duration and bleed level.

Excessive Force is another gauntlet enchantment which adds knockback and inflicts slowness as a sort of stun effect.

Extended Reach is a pike enchantment which increases the block reach distance by 0.5.

Sharpened Head is another pike enchantment which is equivalent to the Sharpness enchant.

Regenerative Assault is an enchantment which can be applied to weapons or tools which returns 10% of damage (per level) dealt as health to the user.

Cleaned up code for pikes and gauntlets.

Fixed particles not showing up for the bleed effect.