Add pikes for endgame tiers
Add pikes for Molten, Tesla, Ventus, Astral, and Starstorm.

Make a model template for pikes and datagen all variations.

Add nuggets for Tesla and Astral ingots, and a shard for Starstorm.

Improve Starstorm Crystal crushing mechanic for better stability. It now drops 2-4 shards, which are used in a crafting recipe to form ingots.

Change sprite orientation for animated blocks from horizontal to vertical.