Tweak Skygazer and Evil Eye
The Skygazer has a new trade option. If the player has an enchanting book, it will offer to use the enchanting book on an enchantable item found in the inventory. It is only available if the book's enchantments are not on the selected item or are of a higher level than is currently on the item.

Make the slowness effect a low tier debuff and make levitation a high tier debuff. Also tweaks targeting conditions to ignore line of sight testing when summoned by a staff.