Buff Ventus Armor
Add a new effect to Ventus armor to make it more worthwhile to use. Along with its jump boost and slow falling effects, it can now reflect incoming projectiles for a short period of time. Like the Astral armor set bonus, there is a cooldown for Ventus armor.

The cooldown is 120 ticks (6 seconds), and the reflection window lasts for 60 ticks (3 seconds). Any incoming projectiles are moved in front of the player's facing direction and launched with an extra 50% velocity. This effectively allows redirection of projectiles.

Bump dependency versions and bump version in update.json.