Rebalance molten, texture updates, IWCB
Rebalance molten with a new smithing template. This requires netherite to be obtained before molten, as netherite gear is used as a base. The template itself can be crafted with molten ingots and obsidian.

Re-texture endgame tiers of gauntlets and pikes.

Enable IWCB bridge with a snapshot version of IWCB.