Improve knockback resistance effects
Knockback resistance (such as through the netherite shield accessory) is now handled with an attribute modifier rather than an event for better compatibility (for example, with the warden's sonic boom, as it does not fire the KB event).

Update debug overlay to display the last damage taken next to the last damage dealt.

Minor loot table updates, added names to pools but also tweaked drop rates of some items.