Add two new accessories, GLM rewrites
Added two new accessories, the Depth Charm and the Insomnia Amulet.

The depth charm provides damage resistance, a general damage increase, and melee knockback. This effect scales with depth, beginning at y<64 (it has no effect above this). The deeper the player goes, the better the stats. It can be found in chests in the deep dark.

The insomnia amulet provides a damage bonus, scaling with the player's insomnia. After one full day/night cycle, the effect begins to build. This maximizes after 7 days, however the player may begin to be randomly effected with an array of debuffs to encourage sleep. It is a drop from phantoms.

Restructure global loot modifiers for simplicity and to reduce redundancy.