Update branding, build restructuring
Update logo files for a more modern look, and add a new cover logo to the README (credit to `@shinyflareon` on Discord).

Fix Tiltros sky rendering without the texture and appearing as full black.

Setup publishing to CurseForge and Modrinth with Gradle.

Create a RELEASE_NOTES.md which contains the changes for the current release, as it would appear on a GitHub release. This will be used for changelogs on CurseForge/Modrinth as well.

Bump dependency versions.