Implement Champion Towers
Implement Champion Tower structures, which are designed to be challenging for players. Each floor contains various mobs with unique attributes and weapons, increasing in difficulty with each level. At the top is Super Hans, which is invulnerable until the rest of the tower minibosses are destroyed. There is plenty of high level loot available throughout the tower.

Fix Champion Keycard Bricks not breaking nearby instances when the player unlocks them in survival mode.

Add loot table to Super Hans. There are three pools, the first two roll multiple times depending on the number of players that dealt damage, the third is simply potion drops and is not affected by the number of players.

Prevent the Wandering Warrior from setting equipment items if the main hand already contains something.

Add mixin into the Entity class that will load passengers from NBT, as for some reason vanilla does not do this.

Add mixin into the Slime class that makes finalizeSpawn respect NBT sizes (MC-87578), and prevents attributes from being reset when spawning from a structure.

Include Debug Utils in the dev environment and switch the Mixin AP from stock to the Fabric fork.