Some general bugfixes / tweaks
Fix tooltip coloring for max level enchants not working before visiting a Skygazer, or being in the vicinity of one.

Rotate Champion Base block texture to be vertical, for compatibility with mods like Rubidium.

Tweak weighting and drops in loot tables for Super Hans and chests inside Champion Towers.

Make loot pool rolls for Celestial Towers be determined by the number of attacking players.

Update advancements and add some that were missing for stuff in the last release.