Implement powder system
Firearms now require a form of powder in order to be used. Either gunpowder, blackpowder, or sulfur dust can be used, each with their own performance values.

Add Potassium Nitrate ore, with potassium nitrate being used in crafting gunpowder and blackpowder.

Increase generation rate of sulfur in the overworld.

Firearms now have a 90% chance to misfire underwater. Additionally, rain and being in powdered snow will reduce the effectiveness of them. Equipping the Powder Horn accessory reduces the impact of this.