Accessory tooltip improvements
Add a new alt-tooltip section for extended information, which lists all of the effects of accessories.

Add a Reinforced Depth Charm item which additionally provides sonic boom resistance.

Add a looting level effect type which is used by the Amethyst Ring, instead of hardcoding it.

Change default range of the Sculk Staff from 25 to 15 blocks.

Change Ventus Staff advancement from a goal to a challenge, to match the other staff advancements.