Improve smoke grenade mechanics
Allow smoke grenades to bounce off walls and entities. When they fall below a certain velocity, smoke will spawn.

Smoke grenades can now be charged when throwing, similar to bows. A higher charge results in a higher throw velocity.

Double the hitbox size of smoke grenades from 0.25 to 0.5 blocks.

Improve renderer so that smoke grenades appear to rotate in the air and will lay on their sides when on the ground.